Add a --tmp-write-capacity option when restoring tables
This option allows to temporary set the table and its indexes
write capacity to a higher value for the time of the restoration,
then reset it to the expected value. Moreover, this temporary
value is only set if the table or the index write capacity is
lower than the temporary one. This allows for faster restoration
with less throttling.

Signed-off-by: Raphaël Beamonte <raphael.beamonte@bhvr.com>